질문 응답 식별자를 메시지 ID로 맞춘다
Question은 답변 메시지로 생성할 수 없도록 막고, IS-001 응답 questionId를 turn이 아니라 저장된 Message.id로 반환해 questions.message_id 키 설계와 API 응답을 맞춘다.

Constraint: turn은 Message 저장 순서/세션 내 위치로 유지하되 외부 응답 식별자는 Message.id를 사용한다.

Rejected: questionId를 turn 1~3으로 계속 반환 | questions 테이블의 단일 messageId PK/FK 설계와 어긋남.

Confidence: high

Scope-risk: narrow

Directive: Question 생성 시 ANSWER 역할 Message를 허용하지 말고, 응답 questionId는 영속화된 Message.id로 유지한다.

Tested: .\\gradlew.bat test --tests com.careerdungeon.domain.interview.controller.InterviewControllerIntegrationTest --tests com.careerdungeon.domain.interview.repository.InterviewPersistenceRepositoryTest

Not-tested: 전체 테스트 재실행은 직전 변경 중 통과 확인, 최종 커밋 직전에는 관련 테스트만 재실행

+++ src/main/java/com/careerdungeon/domain/interview/dto/InterviewQuestionResponse.java
@@ -1,7 +1,7 @@
 package com.careerdungeon.domain.interview.dto;
 
 public record InterviewQuestionResponse(
-        int questionId,
+        Long questionId,
         String question
 ) {
 }
